chore(GH-1472): move ErrorMessages within zmscitizenapi and update usages
