+++ zmscitizenapi/src/Zmscitizenapi/Middleware/RateLimitingMiddleware.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace BO\Zmscitizenapi\Middleware;
 
 use BO\Zmscitizenapi\Helper\ClientIpHelper;
-use BO\Zmscitizenapi\Localization\ErrorMessages;
+use BO\Zmscitizenapi\Helper\ErrorMessages;
 use BO\Zmscitizenapi\Services\Core\LoggerService;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
