+++ zmscitizenapi/src/Zmscitizenapi/Services/Core/ExceptionService.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace BO\Zmscitizenapi\Services\Core;
 
-use BO\Zmscitizenapi\Localization\ErrorMessages;
+use BO\Zmscitizenapi\Helper\ErrorMessages;
 
 class ExceptionService
 {
