+++ zmscitizenapi/tests/Zmscitizenapi/Controllers/Availability/AvailableAppointmentsListControllerTest.php
@@ -2,7 +2,7 @@
 
 namespace BO\Zmscitizenapi\Tests\Controllers\Availability;
 
-use BO\Zmscitizenapi\Localization\ErrorMessages;
+use BO\Zmscitizenapi\Helper\ErrorMessages;
 use BO\Zmscitizenapi\Tests\ControllerTestCase;
 
 class AvailableAppointmentsListControllerTest extends ControllerTestCase
