+++ zmscitizenapi/tests/Zmscitizenapi/Controllers/Office/OfficesListControllerTest.php
@@ -2,7 +2,7 @@
 
 namespace BO\Zmscitizenapi\Tests\Controllers\Office;
 
-use BO\Zmscitizenapi\Localization\ErrorMessages;
+use BO\Zmscitizenapi\Helper\ErrorMessages;
 use BO\Zmscitizenapi\Tests\ControllerTestCase;
 class OfficesListControllerTest extends ControllerTestCase
 {
