+++ zmscitizenapi/tests/Zmscitizenapi/Controllers/Service/ServiceListByOfficeControllerTest.php
@@ -2,7 +2,7 @@
 
 namespace BO\Zmscitizenapi\Tests\Controllers\Service;
 
-use BO\Zmscitizenapi\Localization\ErrorMessages;
+use BO\Zmscitizenapi\Helper\ErrorMessages;
 use BO\Zmscitizenapi\Tests\ControllerTestCase;
 
 class ServiceListByOfficeControllerTest extends ControllerTestCase
