+++ zmscitizenapi/tests/Zmscitizenapi/Services/Core/ExceptionServiceTest.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace BO\Zmscitizenapi\Tests\Services\Core;
 
-use BO\Zmscitizenapi\Localization\ErrorMessages;
+use BO\Zmscitizenapi\Helper\ErrorMessages;
 use BO\Zmscitizenapi\Services\Core\ExceptionService;
 use PHPUnit\Framework\TestCase;
